backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0.post1
pydantic==2.5.2
langchain==0.1.16
langchain-groq==0.1.3
langchain-google-genai==1.0.3
langgraph==0.0.38
python-dotenv==1.0.0
docker==7.1.0
requests==2.31.0
urllib3==1.26.18
anyio==3.7.1
pillow==10.3.0